requirements.txt
# Core dependencies
numpy==2.1.*
pyaudio==0.2.14
pyyaml==6.0.2
requests==2.31.0

# Text-to-speech
piper-tts==1.3.0

# Speech-to-text
openai-whisper
